+++ includes/Achievement.php
@@ -19,7 +19,7 @@ class Achievement {
 	/**
 	 * @param array $info arguments:
 	 * key:
-	 * user: The user who earned the achievement;
+	 * user: The user who earned the achievement.
 	 */
 	public static function achieve( $info ) {
 		if ( empty( $info['key'] ) ) {
@@ -32,15 +32,110 @@ class Achievement {
 		}
 		$key = $info['key'];
 		$config = MediaWikiServices::getInstance()->getMainConfig();
+		$registry = $config->get( Constants::ACHIEVEMENT_BADGES_ACHIEVEMENTS );
 
-		$registry = $config->get( Constants::CONFIG_KEY_ACHIEVEMENT_BADGES_ACHIEVEMENTS );
 		if ( !isset( $registry[$key] ) ) {
 			throw new MWException( "Achievement key not found: {$key}" );
 		}
+		$registry[$key]['type'] = $registry[$key]['type'] ?? 'instant';
+		if ( $registry[$key]['type'] !== 'instant' ) {
+			throw new MWException(
+				"$__METHOD__ is called with only an instant achievement, but $key is not" );
+		}
 
+		$count = self::selectLogCount( $key, $user );
+		if ( $count > 0 ) {
+			// The achievement was earned already.
+			return;
+		}
+
+		self::achieveInternal( $key, $user );
+	}
+
+	/**
+	 * @param array $info arguments:
+	 * key:
+	 * user: The user who earned the achievement.
+	 * stats:
+	 */
+	public static function sendStats( $info ) {
+		if ( empty( $info['key'] ) ) {
+			throw new MWException( "'key' parameter is mandatory" );
+		}
+
+		$user = $info['user'];
+		if ( !self::isAchievementBadgesAvailable( $user ) ) {
+			wfDebug( '[AchievementBadges] user cannot use AchievementBadges' );
+			return;
+		}
+		$key = $info['key'];
+		$stats = $info['stats'];
+		$config = MediaWikiServices::getInstance()->getMainConfig();
+		$registry = $config->get( Constants::ACHIEVEMENT_BADGES_ACHIEVEMENTS );
+		if ( !isset( $registry[$key] ) ) {
+			throw new MWException( "Achievement key not found: {$key}" );
+		}
+		if ( $registry[$key]['type'] !== 'stats' ) {
+			throw new MWException( "Only instant achievement can be called by " . __METHOD__ );
+		}
+		$thresholds = $registry[$key]['thresholds'];
+		wfDebug( "[AchievementBadges] Check {$stats} is in thresholds " . implode( $thresholds, ', ' ) );
+		if ( $stats < $thresholds[0] ) {
+			return;
+		}
+		$numThresholds = count( $thresholds );
+
+		$earned = self::selectLogCount( $key, $user );
+		wfDebug( '[AchievementBadges] ' . $user->getName() . " has $earned achieved achievements" );
+		if ( $earned == $numThresholds ) {
+			return;
+		}
+
+		for ( $i = $earned; $i < $numThresholds; $i++ ) {
+			if ( $stats >= $thresholds[$i] ) {
+				self::achieveInternal( $key, $user, $i );
+			} else {
+				break;
+			}
+		}
+	}
+
+	/**
+	 * @param string $key
+	 * @param User $user
+	 * @param int|null $index
+	 */
+	private static function achieveInternal( $key, User $user, $index = null ) {
+		wfDebug( '[AchievementBadges] ' . $user->getName() . ' obtained ' . $key .
+			( $index ? " with $index" : '' ) );
+
+		$suffixedKey = $key;
+		if ( $index && $index > 0 ) {
+			$suffixedKey .= (string)( $index + 1 );
+		}
+		$logEntry = new ManualLogEntry( Constants::LOG_TYPE, $key );
+		$logEntry->setPerformer( $user );
+		$logEntry->setTarget( SpecialPage::getTitleFor( SpecialAchievements::PAGE_NAME ) );
+		$logEntry->setParameters( [
+			'4::key' => $key,
+			'index' => $index,
+		] );
+		$logEntry->insert();
+
+		EchoEvent::create( [
+			'type' => Constants::EVENT_KEY_EARN,
+			'agent' => $user,
+			'extra' => [ 'key' => $suffixedKey ],
+		] );
+	}
+
+	/**
+	 * @param string $key
+	 * @param User $user
+	 */
+	private static function selectLogCount( $key, User $user ) {
 		$dbr = wfGetDB( DB_REPLICA );
-		$query = [];
-		$count = $dbr->selectRowCount(
+		return $dbr->selectRowCount(
 			[ 'logging', 'actor' ],
 			'*',
 			[
@@ -54,28 +149,6 @@ class Achievement {
 				'actor' => [ 'JOIN', 'actor_id = log_actor' ],
 			]
 		);
-		if ( $count ) {
-			// The achievement was earned already.
-			return;
-		}
-
-		wfDebug( '[AchievementBadges] ' . $user->getName() . ' obtained ' . $key );
-
-		$logEntry = new ManualLogEntry( Constants::LOG_TYPE, $key );
-		$logEntry->setPerformer( $user );
-		$logEntry->setTarget( SpecialPage::getTitleFor( SpecialAchievements::PAGE_NAME ) );
-		$logEntry->setParameters( [
-			'4::key' => $key,
-		] );
-		$logEntry->insert();
-
-		$result = EchoEvent::create( [
-			'type' => Constants::EVENT_KEY_EARN,
-			'extra' => [
-				'key' => $key,
-			],
-			'agent' => $user,
-		] );
 	}
 
 	/**