+++ includes/HookHandler/AchievementRegister.php
@@ -4,12 +4,15 @@ namespace MediaWiki\Extension\AchievementBadges\HookHandler;
 
 use Config;
 use ExtensionRegistry;
+use MediaWiki\Extension\AchievementBadges\Achievement;
+use MediaWiki\Extension\AchievementBadges\Constants;
 use MediaWiki\MediaWikiServices;
 use User;
 
 class AchievementRegister implements
-				\MediaWiki\User\Hook\UserSaveSettingsHook,
-				\MediaWiki\Hook\AddNewAccountHook
+				\MediaWiki\Hook\AddNewAccountHook,
+				\MediaWiki\Storage\Hook\PageSaveCompleteHook,
+				\MediaWiki\User\Hook\UserSaveSettingsHook
 	{
 
 	/**
@@ -32,17 +35,26 @@ class AchievementRegister implements
 		if ( $config->get( Constants::CONFIG_KEY_ENABLE_BETA_FEATURE )
 			&& ExtensionRegistry::getInstance()->isLoaded( 'BetaFeatures' ) ) {
 			$achievements[Constants::ACHV_KEY_ENABLE_ACHIEVEMENT_BADGES] = [
+				'type' => 'instant',
+				'checker' => 'MediaWiki\Extension\AchievementBadges\AchievementChecker::checkAlwaysTrue',
 				'priority' => 0,
 				'icon' => '',
-				'checker' =>
-					'MediaWiki\Extension\AchievementBadges\AchievementChecker::checkAlwaysTrue',
 			];
 		} else {
 			$achievements[Constants::ACHV_KEY_SIGN_UP] = [
+				'type' => 'instant',
+				'checker' => 'MediaWiki\Extension\AchievementBadges\AchievementChecker::checkAlwaysTrue',
 				'priority' => 0,
 				'icon' => '',
-				'checker' =>
-					'MediaWiki\Extension\AchievementBadges\AchievementChecker::checkAlwaysTrue',
+			];
+		}
+
+		if ( $config->get( Constants::CONFIG_KEY_REPLACE_ECHO_THANK_YOU_EDIT ) ) {
+			$achievements[Constants::ACHV_KEY_EDIT_COUNT] = [
+				'type' => 'stats',
+				'thresholds' => [ 1, 10, 100, 1000, 10000 ],
+				'priority' => 100,
+				'icon' => '',
 			];
 		}
 	}
@@ -54,15 +66,23 @@ class AchievementRegister implements
 		if ( $user->isAnon() ) {
 			return;
 		}
+		$config = MediaWikiServices::getInstance()->getMainConfig();
+		$registry = $config->get( Constants::ACHIEVEMENT_BADGES_ACHIEVEMENTS );
+		if ( !isset( $registry[Constants::ACHV_KEY_SIGN_UP] ) ) {
+			return;
+		}
 		Achievement::achieve( [ 'key' => Constants::ACHV_KEY_SIGN_UP, 'user' => $user ] );
 	}
 
 	/**
-	 * @param User $user the User object that was created.
-	 * @param bool $byEmail true when account was created "by email"
-	 * @return bool|void True or no return value to continue or false to abort
+	 * @inheritDoc
 	 */
 	public function onAddNewAccount( $user, $byEmail ) {
+		$config = $this->config;
+		$registry = $config->get( Constants::ACHIEVEMENT_BADGES_ACHIEVEMENTS );
+		if ( !isset( $registry[Constants::ACHV_KEY_SIGN_UP] ) ) {
+			return;
+		}
 		Achievement::achieve( [ 'key' => Constants::ACHV_KEY_SIGN_UP, 'user' => $user ] );
 	}
 
@@ -70,7 +90,7 @@ class AchievementRegister implements
 	 * @inheritDoc
 	 */
 	public function onUserSaveSettings( $user ) {
-		if ( !$this->config->get( Constants::CONFIG_KEY_ACHIEVEMENT_BADGES_ENABLE_BETA_FEATURE ) ) {
+		if ( !$this->config->get( Constants::CONFIG_KEY_ENABLE_BETA_FEATURE ) ) {
 			return true;
 		}
 		if ( $user->getOption( Constants::PREF_KEY_ACHIEVEMENT_ENABLE ) ) {
@@ -79,5 +99,33 @@ class AchievementRegister implements
 				'user' => $user,
 			] );
 		}
+	}
+
+	/**
+	 * @inheritDoc
+	 */
+	public function onPageSaveComplete(
+		$wikiPage,
+		$user,
+		$summary,
+		$flags,
+		$revisionRecord,
+		$editResult
+	) {
+		if ( $editResult->isNullEdit() ) {
+			wfDebug( '[AchievementBadges] null edit is ignored.' );
+			return;
+		}
+		$user = User::newFromIdentity( $user );
+		if ( $user->isAnon() ) {
+			return;
+		}
+
+		$editCount = $user->getEditCount() + 1;
+		Achievement::sendStats( [
+			'key' => Constants::ACHV_KEY_EDIT_COUNT,
+			'user' => $user,
+			'stats' => $editCount,
+		] );
 	}
 }